requirements-dev.txt
black==23.3.0
django-extra-checks==0.13.3
django-querycount==0.8.3
django_migration_linter==4.1.0
django_test_migrations==1.3
flake8==6
isort==5.12
nplusone==1
pre-commit==3.3.2
pytest==7.3.1
